Fish.py
Progress! I made a subclass of Fish called leftPiranha and I finally
got it so the piranha swims to the left. I tried to get it to reset
back to the right side of the screen when it leaves on the left, but
the new resetpos(self) function I wrote just for the subclass isn’t
working for some reason.